.initial upload
Image roulette custom scripts - With this you can view the attachments
(images) of connected entities (like service activities or appointments)
inside of an web resource in the Contact entity (they are linked).

It has bulgarian interface, but this can be fixed easily in the html.

It also has modes for comparing images from different appointments or to
just see all of the images grouped by an Appointment.